feat: Moved common code to BaseRuleEmailValidator. Fixed EmailFormatRule error
