+++ app/src/Domain/ValueObjects/Email/Rules/EmailFormatRule.php
@@ -2,21 +2,15 @@
 
 namespace Pryaniki\App\Domain\ValueObjects\Email\Rules;
 
-use Pryaniki\App\Domain\ValueObjects\Email\ValidationResult;
-
 class EmailFormatRule extends BaseRuleEmailValidator
 {
-
-    public function validate(string $email): ValidationResult
+    protected function applyRule(string $email): bool
     {
-        $isValid = filter_var($email, FILTER_VALIDATE_EMAIL) === false;
+        return filter_var($email, FILTER_VALIDATE_EMAIL) !== false;
+    }
 
-        if (!$isValid) {
-            $this->validationResult->addError('Field email is not valid');
-        }
-
-        $this->validationResult->setIsValid($isValid);
-
-        return $this->validationResult;
+    protected function getValidationErrorMessage(string $email): string
+    {
+        return 'Field email is not valid';
     }
 }
